commit1

+++ task_1.java
@@ -3,11 +3,11 @@ import java.util.Scanner;
 public class task_1 {
     public static int price;
     public static void main(String[] args) {
-        int gasPrice = 43;
+        float gasPrice = 43;
         System.out.println("Какое количество литров бензина Вам необходимо? :");
         Scanner in = new Scanner(System.in);
-        price = (int) in.nextDouble();
-        int total = (int) (price*gasPrice);
+        price = in.nextInt();
+        float total = (price*gasPrice);
         System.out.println("С Вас: " + total+"руб.");
 
     }
